move navigation to reasonable package
