+++ android/app/src/main/java/com/thebluefolderproject/leafbyte/LeafByteActivity.kt
@@ -12,7 +12,6 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
-import com.thebluefolderproject.leafbyte.activity.LeafByteNavigation
 import com.thebluefolderproject.leafbyte.activity.ui.theme.LeafByteTheme
 import com.thebluefolderproject.leafbyte.google.signin.isGoogleSignInConfigured
 import com.thebluefolderproject.leafbyte.utils.Text
