+++ android/app/src/main/java/com/thebluefolderproject/leafbyte/compose/MainMenuScreen.kt
@@ -38,7 +38,7 @@ import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.thebluefolderproject.leafbyte.R
-import com.thebluefolderproject.leafbyte.activity.LeafByteNavKey
+import com.thebluefolderproject.leafbyte.LeafByteNavKey
 import com.thebluefolderproject.leafbyte.settings.DataStoreBackedSettings
 import com.thebluefolderproject.leafbyte.settings.MockSettings
 import com.thebluefolderproject.leafbyte.settings.SaveLocation
