+++ android/app/src/main/java/com/thebluefolderproject/leafbyte/utils/CameraUtils.kt
@@ -19,7 +19,7 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.platform.LocalContext
 import androidx.core.content.FileProvider
 import com.thebluefolderproject.leafbyte.R
-import com.thebluefolderproject.leafbyte.activity.LeafByteNavKey
+import com.thebluefolderproject.leafbyte.LeafByteNavKey
 import com.thebluefolderproject.leafbyte.compose.MainMenuAlertType
 import java.io.File
