+++ android/app/src/main/java/com/thebluefolderproject/leafbyte/utils/GalleryUtils.kt
@@ -13,7 +13,7 @@ import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContract
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.snapshots.SnapshotStateList
-import com.thebluefolderproject.leafbyte.activity.LeafByteNavKey
+import com.thebluefolderproject.leafbyte.LeafByteNavKey
 import com.thebluefolderproject.leafbyte.compose.MainMenuAlertType
 
 @Composable
